+++ src/app/page.tsx
@@ -1,3 +1,4 @@
+import { ExpertSection } from "@/components/home/ExpertSection";
 import { HeroSection } from "@/components/home/HeroSection";
 import { LearningJourneySection } from "@/components/home/LearningJourneySection";
 import { ProgramsSection } from "@/components/home/ProgramsSection";
@@ -10,6 +11,7 @@ export default function Home() {
       <TrustSection />
       <ProgramsSection />
       <LearningJourneySection />
+      <ExpertSection />
     </>
   );
 }